% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPRclass.R
\name{cov_func}
\alias{cov_func}
\title{Covariance functions}
\usage{
\preformatted{cov_func(func, ...)}
}
\arguments{
\item{func}{A covariance function, the parameters of which are to be set}

\item{...}{arguments passed on to \code{func}}
}
\value{
a covariance function with fixed parameters
}
\description{
Creates suitable covariance functions for the constructor of \code{GPR} and
\code{GPC} or for creating covariance matrices.
}
\examples{
 cov <- cov_func(sqrexp, l = 0.1)
 cov <- cov_func(rationalquadratic, l = 1, alpha = 0.5)

}
