% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{examples}
\alias{examples}
\title{Example plots of Gaussian Process simulations}
\usage{
\preformatted{examples()}
}
\value{
Nothing
}
\description{
Runs examples of \code{simulate_regression}, \code{simulate_regression_gp} and
\code{simulate_classification}
}
