% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianprocess.R
\docType{package}
\name{gaussianprocess-package}
\alias{gaussianprocess}
\alias{gaussianprocess-package}
\title{Gaussian Process Regression}
\description{
This package implements Gaussian Process Regression and Classification
}
\details{
gaussianprocess implements regression an classification functions for a Gaussian Process
}
\author{
\strong{Maintainer}: Moritz Haas \email{ha.moritz@gmail.com}

Authors:
\itemize{
  \item Michael Budjan \email{m.budjan@yahoo.de}
  \item Konstantin Klumpp
  \item Tim Reitze
}

}
\keyword{internal}
