% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gprc.R
\docType{package}
\name{gprc-package}
\alias{gprc}
\alias{gprc-package}
\title{Gaussian Process Regression}
\description{
This package implements methods for plotting and optimizing Gaussian Process Regression and Classification
}
\details{
gprc implements methods for plotting and optimizing Gaussian process regression and classification
}
\author{
\strong{Maintainer}: Moritz Haas \email{moritz.haas@stud.uni-heidelberg.de}

Authors:
\itemize{
  \item Michael Budjan \email{m.budjan@yahoo.de}
  \item Konstantin Klumpp
  \item Tim Reitze
}

}
\keyword{internal}
