% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{iid_noise}
\alias{iid_noise}
\title{Independent and identically distributed noise}
\usage{
\preformatted{iid_noise(distribution, ...)}
}
\arguments{
\item{distribution}{A function generating random numbers}

\item{...}{arguments passed on to \code{distribution}}
}
\value{
a suitable noise function for \code{simulate_regression} and
\code{simulate_regression_gp}
}
\description{
Creates suitable noise functions for \code{simulate_regression} and
\code{simulate_regression_gp} with iid noise generated by standard
distributions.
}
\details{
The first argument of \code{distribution} has to be the number of random
numbers to generate.
}
\examples{
 observation_noise <- iid(rnorm, sd = 0.1)
 observation_noise <- iid(rcauchy)

}
