% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{Simulation}
\alias{Simulation}
\title{Simulation for Regression and Classification}
\usage{
\preformatted{simulate_regression(func, limits, training_points, num_data = 10, 
noise = 0, error = function(x) 0, cov_names = names(cov_dict))

simulate_classification(func, training_points, limits, k, num_data = 10)
}
}
\description{
Simulates data for Regression and Classification problems, which can then be analyzed via a Gaussian process
}
\section{Functions}{
NA
}

\examples{

}
