% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\docType{data}
\name{fit}
\alias{fit}
\alias{cov_dict}
\title{Optimization of hyperparameters}
\format{An object of class \code{list} of length 6.}
\usage{
cov_dict
}
\description{
Applies optimization methods to find optimal parameters of the covariance function for a given set of datapoints
}
\section{Usage}{
 
\preformatted{fit(X, y, noise, cov_list)}
}

\section{Arguments}{


  \code{X} matrix of inputs

  \code{y} numeric vector of targets

  \code{noise} the inflicted noise of the observations

  \code{cov_list} a list of names of covariance functions
}

\references{
Rasmussen, Carl E. W., Christopher K. I. (2006).	Gaussian processes for machine learning
}
\keyword{datasets}
