% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Optimization of hyperparameters}
\description{
Applies optimization methods to find optimal parameters of the covariance function for a given set of datapoints
}
\section{Usage}{
 
\preformatted{fit(X, y, noise, cov_names)}
}

\section{Arguments}{


  \code{X} matrix of inputs

  \code{y} numeric vector of targets

  \code{noise} the inflicted noise of the observations

  \code{cov_names} a list of names of covariance functions
}

\references{
Rasmussen, Carl E.; Williams, Christopher K. I. (2006).	Gaussian processes for machine learning
}
