% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPCclass.R
\name{GPC}
\alias{GPC}
\title{Predictions and Plots for Gaussian process regression}
\usage{
\preformatted{GPC <- GPC$new(X, y, cov_Fun, noise)


GPC$predict(X*)
GPC$plot(testpoints)
}
}
\description{
Implements a gaussian process and gives tools to predict and plot its values for given testpoints
}
\section{Arguments}{


  \code{X} matrix of inputs

  \code{y} numeric vector of targets

  \code{cov_fun} the chosen covariance function of the gaussian process

  \code{noise} the inflicted noise of the observations

  \code{X*} a numeric vector as the test input

  \code{testpoints} a matrix of testpoints
}

\section{Methods}{


\code{$predict()} returns a numeric vector of the expected value of the underlying 
function f and their variance for the test input

\code{$plot()} displays the results of the predict function for all testpoints in a nice plot
}

\examples{
Hier Beispiele einfügen


}
