% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPCclass.R
\name{GPC}
\alias{GPC}
\title{Predictions and Plots for Gaussian process classification}
\usage{
\preformatted{GPC <- GPC$new(X, y, noise, cov_fun)

GPC$predict(X*)
GPC$plot(testpoints)
}
}
\description{
Implements a gaussian process for classification and gives tools 
 to predict and plot its values for given testpoints
}
\section{Arguments}{


  \code{X} matrix of inputs

  \code{y} numeric vector of targets with values in \{-1,1\}

  \code{cov_fun} the chosen covariance function of the gaussian process

  \code{noise} the inflicted noise of the observations

  \code{X*} a numeric vector as the test input

  \code{testpoints} a matrix of testpoints
}

\section{Methods}{


\code{$predict()} returns a numeric vector of the expected value of the underlying 
function f

\code{$plot()} displays the results of the predict function for all testpoints in a nice plot
}

\examples{
Hier Beispiele einfügen


}
\references{
Rasmussen, Carl E.; Williams, Christopher K. I. (2006).	Gaussian processes for machine learning
}
