% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPRclass.R
\docType{data}
\name{GPR}
\alias{GPR}
\title{Predictions and Plots for Gaussian process regression}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GPR
}
\description{
Implements a gaussian process and gives tools to predict and plot its values for given testpoints
}
\section{Usage}{
 
\preformatted{GPR <- GPR$new(X, y, cov_Fun, noise)


GPR$predict(X*)
GPR$plot(testpoints)
}
}

\section{Arguments}{


  \code{X} matrix of inputs

  \code{y} numeric vector of targets

  \code{cov_Fun} the predicted covarianz function of the gaussian process

  \code{noise} the predicted noise of the observations

  \code{X*} a numeric vector as the test input

  \code{testpoints} a matrix of testpoints
}

\section{Methods}{

\code{$predict()} returns a numeric vector of the expected value of the underlying function f and their variance for the test input

\code{$plot()} displays the results of the predict function for all testpoints in a nice plot
}

\examples{
Hier Beispiele einfügen


}
\keyword{datasets}
