% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_classification}
\alias{simulate_classification}
\title{Simulation for Classification}
\usage{
\preformatted{simulate_classification(func, training_points, limits, k, num_data = 10)
}
}
\arguments{
\item{func}{A function for the simulation of data points}

\item{training_points}{A vector of training points}

\item{limits}{}
}
\description{
Simulates data for Classification problems, which can then be analyzed via a Gaussian process
}
\examples{

}
