% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_regression}
\alias{simulate_regression}
\title{Simulation for Regression}
\usage{
\preformatted{simulate_regression(func, limits, training_points, num_data = 10, 
noise = 0, error = function(x) 0, cov_names = names(cov_dict))
}
}
\arguments{
\item{func}{}
}
\description{
Simulates data for Regression problems, which can then be analyzed via a Gaussian process
}
\examples{

}
